Moved methods to `UserAdmin` class.
