+++ lib/Github/Api/User.php
@@ -182,32 +182,4 @@ class User extends AbstractApi
     {
         return $this->get('users/'.rawurlencode($username) . '/events/public');
     }
-
-    /**
-     * Suspend a user
-     *
-     * @link https://developer.github.com/v3/users/administration/#suspend-a-user
-     *
-     * @param string $username
-     *
-     * @return array
-     */
-    public function suspend($username)
-    {
-        return $this->put('users/'.rawurldecode($username).'/suspended', array('Content-Length' => 0));
-    }
-
-    /**
-     * Unsuspend a user
-     *
-     * @link https://developer.github.com/v3/users/administration/#unsuspend-a-user
-     *
-     * @param string $username
-     *
-     * @return array
-     */
-    public function unsuspend($username)
-    {
-        return $this->delete('users/'.rawurldecode($username).'/suspended');
-    }
 }
