+++ test/Github/Tests/Api/Enterprise/UserAdminTest.php
@@ -1,0 +1,43 @@
+<?php
+namespace Github\Tests\Api\Enterprise;
+
+use Github\Tests\Api\TestCase;
+
+class UserAdminTest extends TestCase
+{
+    /**
+     * @test
+     */
+    public function shouldSuspendUser()
+    {
+        $expectedArray = array();
+
+        $api = $this->getApiMock();
+        $api->expects($this->once())
+            ->method('put')
+            ->with('users/l3l0/suspended')
+            ->will($this->returnValue($expectedArray));
+        $this->assertEquals($expectedArray, $api->suspend('l3l0'));
+    }
+
+    /**
+     * @test
+     */
+    public function shouldUnsuspendUser()
+    {
+        $expectedArray = array();
+
+        $api = $this->getApiMock();
+        $api->expects($this->once())
+            ->method('delete')
+            ->with('users/l3l0/suspended')
+            ->will($this->returnValue($expectedArray));
+
+        $this->assertEquals($expectedArray, $api->unsuspend('l3l0'));
+    }
+
+    protected function getApiClass()
+    {
+        return 'Github\Api\Enterprise\UserAdmin';
+    }
+}
